Cleanup & Refactor
